+++ examples/ai-lab-native-example/index.ts
@@ -5,6 +5,6 @@
 import {AppRegistry} from 'react-native';
 // import App from './App';
 import {name as appName} from './app.json';
-import {StorybookUIRoot as App} from 'ai-lab-native-storybook';
+import {StorybookUIRoot as App} from './storybook';
 
 AppRegistry.registerComponent(appName, () => App);
