+++ examples/ai-lab-native-example/metro.config.js
@@ -2,10 +2,10 @@ const {getDefaultConfig} = require('metro-config');
 const path = require('path');
 
 const aiLabNativePath = path.resolve(__dirname, '../../packages/ai-lab-native');
-const aiLabNativeStorybookPath = path.resolve(
-  __dirname,
-  '../../packages/ai-lab-native/storybook',
-);
+// const aiLabNativeStorybookPath = path.resolve(
+//   __dirname,
+//   '../../packages/ai-lab-native/storybook',
+// );
 const reactPath = path.resolve(__dirname, 'node_modules/react');
 const reactNativePath = path.resolve(__dirname, 'node_modules/react-native');
 
@@ -18,7 +18,7 @@ module.exports = (async () => {
       assetExts: [...assetExts, 'bin'],
       extraNodeModules: {
         'ai-lab-native': aiLabNativePath,
-        'ai-lab-native-storybook': aiLabNativeStorybookPath,
+        // 'ai-lab-native-storybook': aiLabNativeStorybookPath,
         react: reactPath,
         'react-native': reactNativePath,
       },
