+++ examples/ai-lab-native-example/package.json
@@ -27,6 +27,8 @@
     "@types/react-native": "^0.65.0",
     "@types/react-test-renderer": "^17.0.1",
     "@react-native-async-storage/async-storage": "^1.15.9",
+    "@storybook/api": "^6.3.10",
+    "@storybook/react-native": "^5.3.25",
     "babel-jest": "^26.6.3",
     "eslint": "^7.14.0",
     "jest": "^26.6.3",
