+++ examples/ai-lab-native-example/storybook/AILabNativeImage.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { storiesOf } from '@storybook/react-native';
-import { AILabNativeImage } from './AILabNativeImage';
+import {storiesOf} from '@storybook/react-native';
+import {AILabNativeImage} from 'ai-lab-native';
 
 storiesOf('AILabNativeImage', module).add('to Storybook', () => (
   <AILabNativeImage source={require('./dinner.jpg')} />
