+++ examples/ai-lab-native-example/storybook/beach.jpeg
