+++ examples/ai-lab-native-example/storybook/cat.jpeg
