+++ packages/ai-lab-native/components/index.ts
@@ -1,0 +1,1 @@
+export * from './AILabNativeImage';
