+++ packages/ai-lab-native/index.tsx
@@ -1,6 +1,1 @@
-import * as React from 'react';
-import { Text } from 'react-native';
-
-export const Thing = () => {
-  return <Text>the snozzberries taste like snozzberries</Text>;
-};
+export * from './components';
