+++ packages/ai-lab-native/package.json
@@ -28,8 +28,6 @@
     "@babel/core": "^7.15.5",
     "@react-native-async-storage/async-storage": "^1.15.9",
     "@react-native-community/eslint-config": "^0.0.5",
-    "@storybook/api": "^6.3.10",
-    "@storybook/react-native": "^5.3.25",
     "@types/core-js": "^2.5.5",
     "@types/react": "^17.0.24",
     "@types/react-native": "^0.65.1",
@@ -55,8 +53,5 @@
     "react-native-fs": "^2.18.0",
     "react-native-unimodules": "^0.14.8",
     "react-native-webview": "^11.14.0"
-  },
-  "resolutions": {
-    "@storybook/api": "^6.3.10"
   }
 }
