+++ packages/ai-lab/components/AILabImage/index.ts
@@ -1,0 +1,1 @@
+export * from './AILabImage';
