+++ packages/ai-lab/index.tsx
@@ -1,5 +1,1 @@
-import * as React from 'react';
-
-export const Thing = () => {
-  return <div>the snozzberries taste like snozzberries</div>;
-};
+export * from './components';
